requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
python-dotenv==1.2.1

langchain==1.1.0
langchain-pinecone==0.2.13
langchain-openai==1.1.0
langchain-community==0.4.1
langchain-text-splitters==1.0.0
langchain-core==1.1.0

pinecone-client==6.0.0
pydantic==2.12.5
pypdf==6.4.0

huggingface-hub>=0.20.0

# CPU-only Machine Learning stack (to avoid GPU CUDA downloads)
sentence-transformers==2.2.2
torch==2.2.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
